% Auto-generated: do not edit by hand
\name{''DashAudioRecorder}

\alias{''DashAudioRecorder}

\title{DashAudioRecorder component}

\description{

}

\usage{
''DashAudioRecorder(id=NULL, audioData=NULL, audioStream=NULL, audioType=NULL,
autoGainControl=NULL, currentVolume=NULL,
echoCancellation=NULL, noiseSuppression=NULL,
recordMode=NULL, streamMode=NULL, visualMode=NULL)
}

\arguments{
\item{id}{Character. }

\item{audioData}{Character. }

\item{audioStream}{Character. }

\item{audioType}{Character. }

\item{autoGainControl}{Logical. }

\item{currentVolume}{Numeric. }

\item{echoCancellation}{Logical. }

\item{noiseSuppression}{Logical. }

\item{recordMode}{Character. }

\item{streamMode}{Logical. }

\item{visualMode}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

