% Auto-generated: do not edit by hand
\name{''DashAudioRecorder}

\alias{''DashAudioRecorder}

\title{DashAudioRecorder component}

\description{

}

\usage{
''DashAudioRecorder(id=NULL, audioData=NULL, audioType=NULL, recordMode=NULL,
visualMode=NULL)
}

\arguments{
\item{id}{Character. }

\item{audioData}{Character. }

\item{audioType}{Character. }

\item{recordMode}{Character. }

\item{visualMode}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

